Move encoder to its own module
